main: allow passing --address to set bind address
Add a new --address option that takes an IP address values (v4 or v6) to
specify exactly what address to listen on.

Internally, a slice of IP objects is used but we currently only support
binding to one of them. Checking for an empty slice is easy and there's
a chance we will want multiple listeners in the future but to keep
things somewhat simple the CLI and listener only allows for one.

The goal of this feature is to eventually support colocation of smb
services on a single node by having one set of smb related services
bind only to a given address and a different set of smb services
bind only to a different address on the same host.

Signed-off-by: John Mulligan <jmulligan@redhat.com>